pyproject.toml
[project]
name = "chatboot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
]
